Remove path from CodeSample
