+++ packages/gitbook/src/components/DocumentView/OpenAPI/style.css
@@ -366,7 +366,7 @@
 }
 
 .openapi-codesample-header {
-    @apply flex flex-row items-center px-3 py-2 border-b border-dark/2 dark:border-light/2;
+    @apply flex flex-row items-center;
 }
 
 .openapi-codesample-header-content {
@@ -467,15 +467,15 @@
 }
 
 .openapi-operation .openapi-tabs-list {
-    @apply flex flex-row gap-3 py-2 px-3 border-b border-dark/2 dark:border-light/2 w-full;
+    @apply flex flex-row gap-3 py-1.5 px-2.5 border-b border-dark/2 dark:border-light/2 w-full;
 }
 
 .openapi-operation .openapi-tabs-tab {
-    @apply hover:bg-light-3 font-mono tabular-nums dark:hover:bg-dark-3 hover:text-primary cursor-pointer transition-all relative text-sm text-dark/7 px-1 dark:text-light/7 border border-transparent rounded;
+    @apply hover:bg-light-3 font-mono tabular-nums dark:hover:bg-dark-3 hover:text-primary cursor-pointer transition-all relative text-[13px] text-dark/7 px-1 dark:text-light/7 border border-transparent rounded;
 }
 
 .openapi-operation .openapi-tabs-tab[aria-selected='true'] {
-    @apply text-primary after:absolute after:-bottom-2.5 after:left-0 after:w-full after:h-px after:bg-primary after:transition-all;
+    @apply text-primary after:absolute after:-bottom-2 after:left-0 after:w-full after:h-px after:bg-primary after:transition-all;
 }
 
 .openapi-operation .openapi-tabs-panel {
