+++ packages/react-openapi/src/OpenAPIDisclosure.tsx
@@ -1,109 +1,44 @@
-import {
-    Button,
-    Disclosure,
-    DisclosureGroup,
-    DisclosurePanel,
-    Heading,
-    Key,
-} from 'react-aria-components';
 import React from 'react';
+import { OpenAPIClientContext, OpenAPIContextProps } from './types';
+import { mergeProps, useButton, useDisclosure, useFocusRing } from 'react-aria';
+import { useDisclosureState } from 'react-stately';
 
 interface Props {
-    groups: {
-        id: string;
-        label: string | React.ReactNode;
-        body: React.ReactNode;
-    }[];
-    icon?: React.ReactNode;
+    context: OpenAPIClientContext;
+    children: React.ReactNode;
+    label?: string;
 }
 
-type DisclosureGroup = {
-    id: string;
-    label: string | React.ReactNode;
-    body: React.ReactNode;
-};
-
-import { mergeProps, useButton, useDisclosure, useFocusRing, useId } from 'react-aria';
-import {
-    DisclosureGroupProps,
-    DisclosureGroupState,
-    useDisclosureGroupState,
-    useDisclosureState,
-} from 'react-stately';
-
-const DisclosureGroupStateContext = React.createContext<DisclosureGroupState | null>(null);
-
-export function OpenAPIDisclosureGroup(props: DisclosureGroupProps & Props) {
-    const { icon, groups } = props;
-
-    let state = useDisclosureGroupState(props);
-
-    return (
-        <DisclosureGroupStateContext.Provider value={state}>
-            {groups.map((group) => (
-                <DisclosureItem icon={icon} key={group.id} group={group} />
-            ))}
-        </DisclosureGroupStateContext.Provider>
-    );
-}
-
-function DisclosureItem(props: { group: DisclosureGroup; icon?: React.ReactNode }) {
-    const { icon, group } = props;
-
-    let defaultId = useId();
-    const id = group.id || defaultId;
-    let groupState = React.useContext(DisclosureGroupStateContext);
-    let isExpanded = groupState?.expandedKeys.has(id) || false;
-    let state = useDisclosureState({
-        isExpanded,
-        onExpandedChange(isExpanded) {
-            if (groupState) {
-                groupState.toggleKey(id);
-            }
-        },
-    });
-
+export function OpenAPIDisclosure({ context, children, label }: Props): JSX.Element {
+    let state = useDisclosureState({});
     let panelRef = React.useRef<HTMLDivElement | null>(null);
     let triggerRef = React.useRef<HTMLButtonElement | null>(null);
-    let isDisabled = groupState?.isDisabled || false;
-    let { buttonProps: triggerProps, panelProps } = useDisclosure(
-        {
-            ...props,
-            isExpanded,
-            isDisabled,
-        },
-        state,
-        panelRef,
-    );
+    let { buttonProps: triggerProps, panelProps } = useDisclosure({}, state, panelRef);
     let { buttonProps } = useButton(triggerProps, triggerRef);
     let { isFocusVisible, focusProps } = useFocusRing();
 
     return (
-        <div className="openapi-disclosure-group">
+        <div className="openapi-disclosure">
             <button
-                slot="trigger"
                 ref={triggerRef}
                 {...mergeProps(buttonProps, focusProps)}
+                slot="trigger"
+                className="openapi-disclosure-trigger"
                 style={{
                     outline: isFocusVisible
-                        ? '2px solid rgb(var(--primary-color-500)/0.4)'
+                        ? '2px solid rgb(var(--primary-color-500) / 0.4)'
                         : 'none',
                 }}
-                className="openapi-disclosure-group-trigger"
             >
-                <div className="openapi-disclosure-group-icon">
-                    {icon || (
-                        <svg viewBox="0 0 24 24" className="openapi-disclosure-group-icon">
-                            <path d="m8.25 4.5 7.5 7.5-7.5 7.5" />
-                        </svg>
-                    )}
-                </div>
-                {group.label}
+                {context.icons.plus}
+                <span>
+                    {`${state.isExpanded ? 'Hide' : 'Show'} ${label ? label : `child attributes`}`}
+                </span>
             </button>
 
             {state.isExpanded && (
-                <div className="openapi-disclosure-group-panel" ref={panelRef} {...panelProps}>
-                    {group.body}
+                <div ref={panelRef} {...panelProps} className="openapi-disclosure-panel">
+                    {children}
                 </div>
             )}
         </div>
